внесены правки в верстку
